requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.81
langchain_sandbox==0.0.6
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.5.2
numpy==2.3.5
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pgvector==0.4.2
protobuf==6.33.2
psycopg2-binary==2.9.11
pyobvector>=0.2.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy>=2.0.0
strictyaml==1.7.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
